Refactor pauli graph converters
